refactor: add conditions to commands' inputs

+++ src/components/base/command-line.js
@@ -24,44 +24,84 @@ export const commandLine = async (rl, line) => {
     const args = parts.slice(1);
     switch (command) {
         case 'up':
+            if (args.length !== 0) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             process.chdir('..');
             break;
         case 'cd':
+            if (args.length !== 1) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await goToFolder(args);
             break;
         case 'ls':
+            if (args.length !== 0) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await printAllFiles();
             break;
         case 'add':
-            if (!isValidFileFormat(args.join(' '))) {
+            if (!isValidFileFormat(args.join(' ')) || args.length !== 1) {
                 console.error(ERROR_INPUT);
                 break;
             }
             await addFile(args);
             break;
         case 'cat':
+            if (args.length !== 1) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await readContentFile(args);
             break;
         case 'rn':
+            if (args.length !== 2) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await renameFile(args);
             break;
         case 'rm':
+            if (args.length !== 1) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await deleteFile(args);
             break;
         case 'cp':
+            if (args.length !== 2) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await copyFile(args);
             break;
         case 'mv':
+            if (args.length !== 2) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await moveFile(args);
             break;
         case 'hash':
+            if (!isValidFileFormat(args.join(' ')) || args.length !== 1) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await calculateHash(args);
             break;
         case 'compress':
+            if (args.length !== 2) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             await compressFile(args);
             break;
         case 'decompress':
-            if (!args[0].includes('.br')) {
+            if (!args[0].includes('.br') || args.length !== 2) {
                 console.error(ERROR_INPUT);
                 break;
             }
@@ -77,6 +117,10 @@ export const commandLine = async (rl, line) => {
             break;
         }
         case '.exit':
+            if (args.length !== 0) {
+                console.error(ERROR_INPUT);
+                break;
+            }
             rl.close();
             break;
         default:
